+++ plugins/qAnimation/qAnimationDlg.h
@@ -38,7 +38,7 @@ public:
 
 protected slots:
 
-	void onFPSChanged(double);
+	void onFPSChanged(int);
 
 	void onTotalTimeChanged(double);
 	void onStepTimeChanged(double);
